move test helpers into package
