+++ test/analysis.py
@@ -4,8 +4,6 @@ Tests focused on the analysis capabilities of turnkey CLI
 
 import os
 import unittest
-from pathlib import Path
-import shutil
 import glob
 import subprocess
 import numpy as np
@@ -13,12 +11,10 @@ from contextlib import redirect_stdout
 from unittest.mock import patch
 import io
 import sys
-import platform
 from turnkeyml.cli.cli import main as turnkeycli
-import turnkeyml.common.labels as labels
 from turnkeyml.parser import parse
 import turnkeyml.common.filesystem as filesystem
-from helpers import common
+import turnkeyml.common.test_helpers as common
 from turnkeyml.analyze.status import Verbosity
 
 try:
